Extract HeaderImage
